==> Performance_benchmarks/SingleCell/Darmanis_dataset/2_Darmanis_DTU.Rmd <==
---
title: "2_Darmanis_DTU"
author: "Jeroen Gilis"
date: "05/11/2020"
output: html_document
---

**In order to run this script (2_Darmanis_DTU.Rmd), the dataset Darmanis_benchmark_datasets_count.Rdata (or, alternatively, Darmanis_benchmark_datasets_scaledTPM.Rdata) is required.** This file can either be generated with the 1_Darmanis_prepare.Rmd script or downloaded from Zenodo.

Here we run the DTU analyses for all 6 methods on all (18) Darmanis benchmark datasets. Note that for DEXSeq and DRIMSeq we only run the datasets with 20 cells in each group, as these methods do not scale to large datasets. NBSplice was omitted as it does not converge on datasets with many zeroes. This code runs approximately 45 minutes hours on a MacBook Pro 2018, processor; 2,3 GHz Quad-Core Intel Core i5, 16GB RAM. Most of this runtime was attributed to the DEXSeq analysis.

**If you do not want to run this script, its output can also be downloaded from Zenodo: Darmanis_DTU_results_count.Rdata (or, alternatively, Darmanis_DTU_results_scaledTPM.Rdata) **

```{r setup, include=FALSE, echo=FALSE}
require("knitr")
opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

# Load DTU methods

Source file that allows running all seven DTU methods that were assessed in this paper: satuRn, DoubleExpSeq, limma diffsplice, edgeRdiffsplice, DEXSeq, DRIMseq and NBSplice.

```{r, message=FALSE, warning=FALSE}
### load libraries
library(edgeR)
library(limma)
library(DEXSeq)
library(DRIMSeq)
library(DoubleExpSeq)
library(NBSplice)
library(satuRn)

source(file="./Performance_benchmarks/DTU_methods.R")
```

# Set up parallel execution

Note that the optimal parameter setting for parallelizing the DTU analyses depends on your specific computing system.

```{r}
if(TRUE) {
    nrCores <- 2

    if(nrCores != 1) {
        doParallel <- TRUE
        doProgress <- 'none'

        registerDoMC(cores = nrCores)
    } else {
        doParallel <- FALSE
        doProgress <- 'text'
    }
}
```

# Run all analyses 

Run the DTU analyses for all 6 methods on all (18) Darmanis benchmark datasets. Note that for DEXSeq and DRIMSeq we only run the datasets with 20 cells in each group, as these methods do not scale to large datasets. NBSplice was omitted as it does not converge on datasets with many zeroes.
This code runs approximately 45 minutes on a MacBook Pro 2018, processor; 2,3 GHz Quad-Core Intel Core i5, 16GB RAM. Most of this runtime was attributed to the DEXSeq analysis.

```{r}
### Load benchmark data
load(file="./Data/Darmanis_benchmark_datasets_count.Rdata")
#load(file="./Data/Darmanis_benchmark_datasets_scaledTPM.Rdata")

DarmanisBenchmarkData <- c(DarmanisBenchmarkLenient,DarmanisBenchmarkStringent)

### Run DTU analyses on benchmark data

print("start satuRn")

tStart <- Sys.time()
suppressWarnings(DarmanisDtuBenchmark_satuRn <- plyr::llply(
    .data = DarmanisBenchmarkData,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- satuRn_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )
        
        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME
        )]

        ### Return result
        return(
            list(
                dtuAnalysis = localRes
            )
        )
    }
))
difftime(Sys.time(), tStart)

print("start edgeR_diffsplice")

tStart <- Sys.time()
DarmanisDtuBenchmark_edgeRdiffsplice <- plyr::llply(
    .data = DarmanisBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- edgeR_diffsplice_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME
        )]

        ### Return result
        return(
            list(
                dtuAnalysis = localRes
            )
        )
    }
)
difftime(Sys.time(), tStart)

print("start DoubleExpSeq")

tStart <- Sys.time()
DarmanisDtuBenchmark_DoubleExpSeq <- plyr::llply(
    .data = DarmanisBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- DoubleExpSeq_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME
        )]

        ### Return result
        return(
            list(
                dtuAnalysis = localRes
            )
        )
    }
)
difftime(Sys.time(), tStart)

print("start limma diffsplice")

tStart <- Sys.time()
DarmanisDtuBenchmark_limmaDiffsplice <- plyr::llply(
    .data = DarmanisBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- limma_diffsplice_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME
        )]

        ### Return result
        return(
            list(
                dtuAnalysis = localRes
            )
        )
    }
)
difftime(Sys.time(), tStart)

print("start DEXSeq")

tStart <- Sys.time()
DarmanisDtuBenchmark_DEXSeq <- plyr::llply(
    .data = DarmanisBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        
        if(ncol(localData$data) > 40) {
            return(NULL)
        }
      
        ### Perform DTU analysis
        localRes <- DEXSeq_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME
        )]

        ### Return result
        return(
            list(
                dtuAnalysis = localRes
            )
        )
    }
)
difftime(Sys.time(), tStart)

print("start DRIMSeq")

tStart <- Sys.time()
DarmanisDtuBenchmark_DRIMSeq <- plyr::llply(
    .data = DarmanisBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
      
        if(ncol(localData$data) > 40) {
            return(NULL)
        }
      
        ### Perform DTU analysis
        localRes <- DRIMSeq_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME
        )]

        ### Return result
        return(
            list(
                dtuAnalysis = localRes
           )
       )
    }
)
difftime(Sys.time(), tStart)

### add method name to list names for easy post-analysis
names(DarmanisDtuBenchmark_satuRn) <- paste0('satuRn_', names(DarmanisDtuBenchmark_satuRn))
names(DarmanisDtuBenchmark_limmaDiffsplice) <- paste0('limma_diffsplice_', names(DarmanisDtuBenchmark_limmaDiffsplice))
names(DarmanisDtuBenchmark_DEXSeq) <- paste0('DEXSeq_', names(DarmanisDtuBenchmark_DEXSeq))
names(DarmanisDtuBenchmark_DRIMSeq) <- paste0('DRIMSeq_', names(DarmanisDtuBenchmark_DRIMSeq))
names(DarmanisDtuBenchmark_edgeRdiffsplice) <- paste0('edgeR_diffsplice_', names(DarmanisDtuBenchmark_edgeRdiffsplice))
names(DarmanisDtuBenchmark_DoubleExpSeq) <- paste0('DoubleExpSeq_', names(DarmanisDtuBenchmark_DoubleExpSeq))

### Save result
save(
    DarmanisDtuBenchmark_satuRn,
    DarmanisDtuBenchmark_limmaDiffsplice,
    DarmanisDtuBenchmark_DEXSeq,
    DarmanisDtuBenchmark_DRIMSeq,
    DarmanisDtuBenchmark_edgeRdiffsplice,
    DarmanisDtuBenchmark_DoubleExpSeq,
    file="./Data/Darmanis_DTU_results_count.Rdata"
)

# save(
#     DarmanisDtuBenchmark_satuRn,
#     DarmanisDtuBenchmark_limmaDiffsplice,
#     DarmanisDtuBenchmark_DEXSeq,
#     DarmanisDtuBenchmark_DRIMSeq,
#     DarmanisDtuBenchmark_edgeRdiffsplice,
#     DarmanisDtuBenchmark_DoubleExpSeq,
#     file="./Data/Darmanis_DTU_results_scaledTPM.Rdata"
# )
```
